--- src/main/java/com/team6/project/domain/board/model/dao/BoardRepository.java
@@ -1,4 +1,0 @@
-package com.team6.project.domain.board.model.dao;
-
-public interface BoardRepository {
-}
