--- src/main/java/com/team6/project/domain/board/model/dto/BoardDTO.java
@@ -1,4 +1,0 @@
-package com.team6.project.domain.board.model.dto;
-
-public class BoardDTO {
-}
