--- src/main/java/com/team6/project/domain/board/model/entity/Board.java
@@ -1,4 +1,0 @@
-package com.team6.project.domain.board.model.entity;
-
-public class Board {
-}
