--- src/main/java/com/team6/project/domain/board/model/service/BoardService.java
@@ -1,4 +1,0 @@
-package com.team6.project.domain.board.model.service;
-
-public class BoardService {
-}
