--- src/main/java/com/team6/project/domain/grade/controller/GradeController.java
@@ -1,4 +1,0 @@
-package com.team6.project.domain.grade.controller;
-
-public class GradeController {
-}
