--- src/main/java/com/team6/project/domain/grade/model/dao/GradeRepository.java
@@ -1,4 +1,0 @@
-package com.team6.project.domain.grade.model.dao;
-
-public interface GradeRepository {
-}
