--- src/main/java/com/team6/project/domain/grade/model/dto/GradeDTO.java
@@ -1,4 +1,0 @@
-package com.team6.project.domain.grade.model.dto;
-
-public class GradeDTO {
-}
