--- src/main/java/com/team6/project/domain/grade/model/entity/Grade.java
@@ -1,4 +1,0 @@
-package com.team6.project.domain.grade.model.entity;
-
-public class Grade {
-}
