--- src/main/java/com/team6/project/domain/grade/model/service/GradeService.java
@@ -1,4 +1,0 @@
-package com.team6.project.domain.grade.model.service;
-
-public class GradeService {
-}
