--- src/main/java/com/team6/project/global/config/IndexController.java
@@ -1,13 +1,0 @@
-package com.team6.project.global.config;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class IndexController {
-
-    @GetMapping(value = {"/", "/main"})
-    public String mainPage() {
-        return "main/main";
-    }
-}
