+++ src/main/java/com/wanted/projectmodule2lms/domain/board/model/dao/BoardRepository.java
@@ -1,0 +1,4 @@
+package com.wanted.projectmodule2lms.domain.board.model.dao;
+
+public interface BoardRepository {
+}
