+++ src/main/java/com/wanted/projectmodule2lms/domain/board/model/dto/BoardDTO.java
@@ -1,0 +1,4 @@
+package com.wanted.projectmodule2lms.domain.board.model.dto;
+
+public class BoardDTO {
+}
