+++ src/main/java/com/wanted/projectmodule2lms/domain/board/model/entity/Board.java
@@ -1,0 +1,4 @@
+package com.wanted.projectmodule2lms.domain.board.model.entity;
+
+public class Board {
+}
