+++ src/main/java/com/wanted/projectmodule2lms/domain/board/model/service/BoardService.java
@@ -1,0 +1,4 @@
+package com.wanted.projectmodule2lms.domain.board.model.service;
+
+public class BoardService {
+}
