+++ src/main/java/com/wanted/projectmodule2lms/domain/grade/controller/GradeController.java
@@ -1,0 +1,4 @@
+package com.wanted.projectmodule2lms.domain.grade.controller;
+
+public class GradeController {
+}
