+++ src/main/java/com/wanted/projectmodule2lms/domain/grade/model/dao/GradeRepository.java
@@ -1,0 +1,4 @@
+package com.wanted.projectmodule2lms.domain.grade.model.dao;
+
+public interface GradeRepository {
+}
