+++ src/main/java/com/wanted/projectmodule2lms/domain/grade/model/dto/GradeDTO.java
@@ -1,0 +1,4 @@
+package com.wanted.projectmodule2lms.domain.grade.model.dto;
+
+public class GradeDTO {
+}
