+++ src/main/java/com/wanted/projectmodule2lms/domain/grade/model/entity/Grade.java
@@ -1,0 +1,4 @@
+package com.wanted.projectmodule2lms.domain.grade.model.entity;
+
+public class Grade {
+}
