+++ src/main/java/com/wanted/projectmodule2lms/domain/grade/model/service/GradeService.java
@@ -1,0 +1,4 @@
+package com.wanted.projectmodule2lms.domain.grade.model.service;
+
+public class GradeService {
+}
