+++ src/main/resources/application.yaml
@@ -4,7 +4,7 @@ spring:
 
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/menudb
+    url: jdbc:mysql://localhost:3306/meat4
     username: wanted
     password: wanted
 
@@ -12,7 +12,7 @@ spring:
     show-sql: true
     database: mysql
     hibernate:
-      ddl-auto: none
+      ddl-auto: update
 
     
     properties:
