--- src/test/java/com/team6/project/ProjectModule2LmsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.team6.project;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProjectModule2LmsApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
