+++ src/test/java/com/wanted/projectmodule2lms/ProjectModule2LmsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.wanted.projectmodule2lms;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ProjectModule2LmsApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
